requirements.txt
duckduckgo-search
openai
groq
streamlit
streamlit-lottie
python-dotenv
requests
folium
streamlit_folium
geopy
fpdf
